hasher/requirements.txt
argon2-cffi
bcrypt
dotenv
py-cpuinfo
psutil
sqlalchemy
resource
psycopg2-binary